Move `jquery-rails` into `hyp.gemspec`.
Where it belongs, rather than having it in `Gemfile`. One notable change
required to do this is to `require `jquery-rails'` when the engine gets
loaded so that all of that gem's assets (notably the jQuery JS file) get
loaded when the mounting application boots up.

Apparently Rails will do this for you when loading dependenices from a
Gemfile, but not for those from a `.gemspec` file.